업그레이드스탯 데이터 추가 및 캐릭터정보 자료형 변경
업그레이드스탯 데이터 추가 및 캐릭터정보 자료형 변경

+++ Project/Assets/Main/Scripts/FrameWork/Controller/Controller/PlayerDataContainer.cs
@@ -7,9 +7,9 @@ namespace GameBerry
     public static class PlayerDataContainer
     {
         public static int Level;
-        public static int Exp;
+        public static double Exp;
 
-        public static int Gold;
+        public static double Gold;
         public static int Dia;
 
         public static int EquipmentSton;
